event_gateway/requirements.txt
aiohttp
aiohttp_sse
aiohttp_cors
pyjwt
wakeonlan

